Pick the activation package by the WooCommerce version under test
`run:activation` set `woocommerce/activation:latest` for every run. The
activation specs drive WooCommerce's admin UI, so their selectors are tied to
the markup of a release: a change in 11.1.0-rc breaks one, and fixing it breaks
the same test on 11.0.1. QIT offers four stable releases plus a prerelease,
regularly spanning two minor lines, so both versions are live at once and one
artifact cannot serve them.

This is the same situation the Core E2E suite was in, reaching it by a
different route — there the test asserts WooCommerce behaviour, here it drives
WooCommerce's UI — so it gets the same answer: a package version per
WooCommerce `major.minor`, resolved from the versions the Manager publishes.

The selection rule moves to SelectsVersionedTestPackage, used by both commands.
It was already duplicated once, in the Manager, for the runs it creates itself;
a third copy would be one too many. A command declares which key it is
published under and what to fall back to, and the rule stays in one place.

The package is still set unconditionally here, as before: this command owns its
package and does not take one from the caller.

Refs QIT-1081.
